requirements.txt
matplotlib
tensorflow
numpy
tensorflow-probability
absl-py